feat: project restructuration + dot builtin
Moved repl to its own folder, and created a folder for the parsing of arguments.
In each builtin, added env structure to arguments (ISSUE: need to cast env to void when it is not used by the function, need to find a way to remove it).

Removed one test for redirection: redir_lessgreat_io_nbr_bad, it used to crash the test suite (add it later when redirection is fully fixed).

Added more tests for dot

Signed-off-by: Sanassi CISSE <sanassi-mory.cisse@epita.fr>